requirements.txt
numpy
pandas
matplotlib
seaborn
pillow
scikit-learn
tqdm
torch
torchvision
torchaudio
#--index-url https://download.pytorch.org/whl/cu121
#pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121